Fix unit tests and go vet errors
